WEBUI-108: ensure unit tests can run
