+++ nuxeo-coldstorage-web/.eslintignore
@@ -1,2 +1,3 @@
 node_modules/
+coverage/
 target/
